replacing altinn-icons with studio-icons
