+++ frontend/jest.config.js
@@ -47,7 +47,7 @@ const config = {
     '^@altinn/process-editor/(.*)': path.join(__dirname, 'packages/process-editor/src/$1'),
     '^@altinn/policy-editor/(.*)': path.join(__dirname, 'packages/policy-editor/src/$1'),
     '^@altinn/studio-components': path.join(__dirname, 'libs/studio-components/src/$1'),
-    '^@altinn/icons': path.join(__dirname, 'libs/altinn-icons/src/$1'),
+    '^@studio/icons': path.join(__dirname, 'libs/studio-icons/src/$1'),
   },
   testRegex: '(\\.(test|spec))\\.(ts|tsx)$',
   moduleFileExtensions: ['ts', 'tsx', 'js'],
