+++ frontend/libs/studio-components/package.json
@@ -8,9 +8,9 @@
     "lint": "eslint \"src/components/**/*.ts*\""
   },
   "dependencies": {
-    "@altinn/icons": "^0.1.0",
     "@digdir/design-system-react": "0.35.0",
     "@digdir/design-system-tokens": "0.10.0",
+    "@studio/icons": "^0.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
