+++ frontend/libs/studio-icons/jest.config.json
