+++ frontend/libs/studio-icons/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@altinn/icons",
+  "name": "@studio/icons",
   "packageManager": "yarn@3.6.4",
   "version": "0.1.0",
   "main": "./src/index.ts",
