+++ frontend/libs/studio-icons/setupTests.ts
