+++ frontend/libs/studio-icons/src/components/Accordion/Accordion.tsx
@@ -6,7 +6,7 @@ import { SvgTemplate } from '../SvgTemplate';
  * @param {IconProps} props the icon props
  * @returns {JSX.Element} the icon as a react component
  * @example
- * import { Accordion } from '@altinn/icons';
+ * import { Accordion } from '@studio/icons';
  */
 export const Accordion = (props: IconProps): JSX.Element => {
   return (
