+++ frontend/libs/studio-icons/src/components/Accordion/index.ts
