+++ frontend/libs/studio-icons/src/components/Checkbox/index.ts
