+++ frontend/libs/studio-icons/src/components/Group/index.ts
