+++ frontend/libs/studio-icons/src/components/Likert/index.ts
