+++ frontend/libs/studio-icons/src/components/LongText/index.ts
