+++ frontend/libs/studio-icons/src/components/NavBar/index.ts
