+++ frontend/libs/studio-icons/src/components/Paragraph/index.ts
