+++ frontend/libs/studio-icons/src/components/RadioButton/RadioButton.tsx
@@ -6,7 +6,7 @@ import { SvgTemplate } from '../SvgTemplate';
  * @param {IconProps} props the icon props
  * @returns {JSX.Element} the icon as a react component
  * @example
- * import { RadioButton } from '@altinn/icons';
+ * import { RadioButton } from '@studio/icons';
  */
 export const RadioButton = (props: IconProps): JSX.Element => {
   return (
