+++ frontend/libs/studio-icons/src/components/Select/index.ts
