+++ frontend/libs/studio-icons/src/components/ShortText/ShortText.tsx
@@ -6,7 +6,7 @@ import { SvgTemplate } from '../SvgTemplate';
  * @param {IconProps} props the icon props
  * @returns {JSX.Element} the icon as a react component
  * @example
- * import { ShortText } from '@altinn/icons';
+ * import { ShortText } from '@studio/icons';
  */
 export const ShortText = ({ ...props }: IconProps): JSX.Element => {
   return (
