+++ frontend/libs/studio-icons/src/components/ShortText/index.ts
