+++ frontend/libs/studio-icons/src/components/SvgTemplate/SvgTemplate.tsx
