+++ frontend/libs/studio-icons/src/components/SvgTemplate/index.ts
