+++ frontend/libs/studio-icons/src/components/Title/Title.tsx
@@ -6,7 +6,7 @@ import { SvgTemplate } from '../SvgTemplate';
  * @param {IconProps} props the icon props
  * @returns {JSX.Element} the icon as a react component
  * @example
- * import { Title } from '@altinn/icons';
+ * import { Title } from '@studio/icons';
  */
 export const Title = (props: IconProps): JSX.Element => {
   return (
