+++ frontend/libs/studio-icons/src/components/Title/index.ts
