+++ frontend/libs/studio-icons/src/components/index.ts
