+++ frontend/libs/studio-icons/src/index.ts
@@ -1,4 +1,4 @@
-// Re-export all icons from @navikt/aksel-icons to make them available for consumers of altinn/icons
+// Re-export all icons from @navikt/aksel-icons to make them available for consumers of studio/icons
 export * from '@navikt/aksel-icons';
 
 export * from './components';
