+++ frontend/libs/studio-icons/src/types/IconProps.ts
