+++ frontend/libs/studio-icons/tsconfig.json
