+++ frontend/packages/ux-editor/package.json
@@ -4,9 +4,9 @@
   "version": "1.0.1",
   "author": "Altinn",
   "dependencies": {
-    "@altinn/icons": "workspace:^",
     "@mui/material": "5.14.16",
     "@reduxjs/toolkit": "1.9.7",
+    "@studio/icons": "workspace:^",
     "axios": "1.6.0",
     "classnames": "2.3.2",
     "react": "18.2.0",
